+++ sdk/monitor/azure-monitor-query/azure/monitor/query/aio/_logs_query_client_async.py
@@ -7,15 +7,16 @@
 
 from datetime import datetime, timedelta
 from typing import Any, Tuple, Union, Sequence, Dict, List, TYPE_CHECKING
-from azure.core.exceptions import HttpResponseError
 from azure.core.tracing.decorator_async import distributed_trace_async
+from azure.core.exceptions import HttpResponseError
 
 from .._generated.aio._monitor_query_client import MonitorQueryClient
 
 from .._generated.models import BatchRequest, QueryBody as LogsQueryBody
-from .._helpers import process_error, construct_iso8601, order_results
-from .._models import LogsQueryResult, LogsBatchQuery
+from .._helpers import construct_iso8601, order_results, process_error, process_prefer
+from .._models import LogsQueryResult, LogsBatchQuery, LogsQueryPartialResult
 from ._helpers_asyc import get_authentication_policy
+from .._exceptions import LogsQueryError
 
 if TYPE_CHECKING:
     from azure.core.credentials_async import AsyncTokenCredential
@@ -31,7 +32,7 @@ class LogsQueryClient(object):
     """
 
     def __init__(self, credential: "AsyncTokenCredential", **kwargs: Any) -> None:
-        self._endpoint = kwargs.pop('endpoint', 'https://api.loganalytics.io/v1')
+        self._endpoint = kwargs.pop("endpoint", "https://api.loganalytics.io/v1")
         self._client = MonitorQueryClient(
             credential=credential,
             authentication_policy=get_authentication_policy(credential),
@@ -41,13 +42,16 @@ class LogsQueryClient(object):
         self._query_op = self._client.query
 
     @distributed_trace_async
-    async def query(
+    async def query_workspace(
         self,
         workspace_id: str,
         query: str,
         *,
-        timespan: Union[timedelta, Tuple[datetime, timedelta], Tuple[datetime, datetime]],
-        **kwargs: Any) -> LogsQueryResult:
+        timespan: Union[
+            timedelta, Tuple[datetime, timedelta], Tuple[datetime, datetime]
+        ],
+        **kwargs: Any
+    ) -> Union[LogsQueryResult, LogsQueryPartialResult]:
         """Execute an Analytics query.
 
         Executes an Analytics query for data.
@@ -58,7 +62,7 @@ class LogsQueryClient(object):
         :param query: The Kusto query. Learn more about the `Kusto query syntax
          <https://docs.microsoft.com/azure/data-explorer/kusto/query/>`_.
         :type query: str
-        :param timespan: The timespan for which to query the data. This can be a timedelta,
+        :param timespan: Required. The timespan for which to query the data. This can be a timedelta,
          a timedelta and a start datetime, or a start datetime/end datetime.
         :type timespan: ~datetime.timedelta or tuple[~datetime.datetime, ~datetime.timedelta]
          or tuple[~datetime.datetime, ~datetime.datetime]
@@ -72,7 +76,7 @@ class LogsQueryClient(object):
          These can be qualified workspace names, workspace Ids or Azure resource Ids.
         :paramtype additional_workspaces: list[str]
         :return: QueryResults, or the result of cls(response)
-        :rtype: ~azure.monitor.query.LogsQueryResult
+        :rtype: ~azure.monitor.query.LogsQueryResult or ~azure.monitor.query.LogsQueryPartialResult
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         timespan = construct_iso8601(timespan)
@@ -81,41 +85,37 @@ class LogsQueryClient(object):
         server_timeout = kwargs.pop("server_timeout", None)
         additional_workspaces = kwargs.pop("additional_workspaces", None)
 
-        prefer = ""
-        if server_timeout:
-            prefer += "wait=" + str(server_timeout)
-        if include_statistics:
-            if len(prefer) > 0:
-                prefer += ","
-            prefer += "include-statistics=true"
-        if include_visualization:
-            if len(prefer) > 0:
-                prefer += ","
-            prefer += "include-render=true"
+        prefer = process_prefer(
+            server_timeout, include_statistics, include_visualization
+        )
 
         body = LogsQueryBody(
-            query=query,
-            timespan=timespan,
-            workspaces=additional_workspaces,
-            **kwargs
+            query=query, timespan=timespan, workspaces=additional_workspaces, **kwargs
         )
 
         try:
-            return LogsQueryResult._from_generated(await self._query_op.execute( # pylint: disable=protected-access
-                workspace_id=workspace_id,
-                body=body,
-                prefer=prefer,
-                **kwargs
-            ))
-        except HttpResponseError as e:
-            process_error(e)
+            generated_response = (
+                await self._query_op.execute(  # pylint: disable=protected-access
+                    workspace_id=workspace_id, body=body, prefer=prefer, **kwargs
+                )
+            )
+        except HttpResponseError as err:
+            process_error(err, LogsQueryError)
+        response = None
+        if not generated_response.error:
+            response = LogsQueryResult._from_generated( # pylint: disable=protected-access
+                generated_response
+            )
+        else:
+            response = LogsQueryPartialResult._from_generated( # pylint: disable=protected-access
+                generated_response, LogsQueryError
+            )
+        return response
 
     @distributed_trace_async
     async def query_batch(
-        self,
-        queries: Union[Sequence[Dict], Sequence[LogsBatchQuery]],
-        **kwargs: Any
-        ) -> List[LogsQueryResult]:
+        self, queries: Union[Sequence[Dict], Sequence[LogsBatchQuery]], **kwargs: Any
+    ) -> List[Union[LogsQueryResult, LogsQueryError, LogsQueryPartialResult]]:
         """Execute a list of analytics queries. Each request can be either a LogQueryRequest
         object or an equivalent serialized model.
 
@@ -124,22 +124,32 @@ class LogsQueryClient(object):
         :param queries: The list of Kusto queries to execute.
         :type queries: list[dict] or list[~azure.monitor.query.LogsBatchQuery]
         :return: list of LogsQueryResult objects, or the result of cls(response)
-        :rtype: list[~azure.monitor.query.LogsQueryResult]
+        :rtype: list[~azure.monitor.query.LogsQueryResult or ~azure.monitor.query.LogsQueryPartialResult
+         or ~azure.monitor.query.LogsQueryError]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         try:
-            queries = [LogsBatchQuery(**q) for q in queries]
+            queries = [LogsBatchQuery(**q) for q in queries]  # type: ignore
         except (KeyError, TypeError):
             pass
-        queries = [q._to_generated() for q in queries] # pylint: disable=protected-access
+        queries = [
+            q._to_generated() for q in queries # pylint: disable=protected-access
+        ]
         try:
             request_order = [req.id for req in queries]
         except AttributeError:
-            request_order = [req['id'] for req in queries]
+            request_order = [req["id"] for req in queries]
         batch = BatchRequest(requests=queries)
         generated = await self._query_op.batch(batch, **kwargs)
         mapping = {item.id: item for item in generated.responses}
-        return order_results(request_order, mapping, LogsQueryResult)
+        return order_results(
+            request_order,
+            mapping,
+            obj=LogsQueryResult,
+            err=LogsQueryError,
+            partial_err=LogsQueryPartialResult,
+            raise_with=LogsQueryError,
+        )
 
     async def __aenter__(self) -> "LogsQueryClient":
         await self._client.__aenter__()