+++ sdk/monitor/azure-monitor-query/samples/sample_batch_query.py
@@ -4,7 +4,7 @@
 from datetime import datetime, timedelta
 import os
 import pandas as pd
-from azure.monitor.query import LogsQueryClient, LogsBatchQuery
+from azure.monitor.query import LogsQueryClient, LogsBatchQuery, LogsQueryStatus, LogsQueryPartialResult
 from azure.identity import DefaultAzureCredential
 
 
@@ -20,27 +20,36 @@ requests = [
         workspace_id= os.environ['LOG_WORKSPACE_ID']
     ),
     LogsBatchQuery(
-        query= """AppRequests | take 10  |
-            summarize avgRequestDuration=avg(DurationMs) by bin(TimeGenerated, 10m), _ResourceId""",
-        timespan=(datetime(2021, 6, 2), timedelta(hours=1)),
+        query= """AppRequestsss | take 10""",
+        timespan=(datetime(2021, 6, 2), timedelta(days=1)),
         workspace_id= os.environ['LOG_WORKSPACE_ID']
     ),
     LogsBatchQuery(
-        query= "AppRequests | take 5",
+        query= """let Weight = 92233720368547758;
+        range x from 1 to 3 step 1
+        | summarize percentilesw(x, Weight * 100, 50)""",
         workspace_id= os.environ['LOG_WORKSPACE_ID'],
         timespan=(datetime(2021, 6, 2), datetime(2021, 6, 3)),
         include_statistics=True
     ),
 ]
-responses = client.query_batch(requests)
+results = client.query_batch(requests)
 
-for response in responses:
-    try:
-        table = response.tables[0]
+for res in results:
+    if res.status == LogsQueryStatus.FAILURE:
+        # this will be a LogsQueryError
+        print(res.message)
+    elif res.status == LogsQueryStatus.PARTIAL:
+        ## this will be a LogsQueryPartialResult
+        print(res.partial_error.message)
+        table = res.tables[0]
         df = pd.DataFrame(table.rows, columns=table.columns)
         print(df)
-        print("\n\n-------------------------\n\n")
-    except TypeError:
-        print(response.error.innererror)
+    elif res.status == LogsQueryStatus.SUCCESS:
+        ## this will be a LogsQueryResult
+        table = res.tables[0]
+        df = pd.DataFrame(table.rows, columns=table.columns)
+        print(df)
+
 
 # [END send_query_batch]