+++ sdk/monitor/azure-monitor-query/samples/sample_server_timeout.py
@@ -11,7 +11,7 @@ credential  = DefaultAzureCredential()
 
 client = LogsQueryClient(credential)
 
-response = client.query(
+response = client.query_workspace(
     os.environ['LOG_WORKSPACE_ID'],
     "range x from 1 to 10000000000 step 1 | count",
     server_timeout=1,