+++ sdk/schemaregistry/azure-schemaregistry/samples/async_samples/sample_code_schemaregistry_async.py
@@ -27,52 +27,52 @@ import os
 import asyncio
 
 from azure.schemaregistry.aio import SchemaRegistryClient
-from azure.schemaregistry import SerializationType
+from azure.schemaregistry import SchemaFormat
 from azure.identity.aio import ClientSecretCredential, DefaultAzureCredential
 
 
 def create_client():
     # [START create_sr_client_async]
-    SCHEMA_REGISTRY_ENDPOINT = os.environ['SCHEMA_REGISTRY_ENDPOINT']
+    SCHEMA_REGISTRY_FQN = os.environ['SCHEMA_REGISTRY_FULLY_QUALIFIED_NAMESPACE']
     token_credential = DefaultAzureCredential()
-    schema_registry_client = SchemaRegistryClient(endpoint=SCHEMA_REGISTRY_ENDPOINT, credential=token_credential)
+    schema_registry_client = SchemaRegistryClient(fully_qualified_namespace=SCHEMA_REGISTRY_FQN, credential=token_credential)
     # [END create_sr_client_async]
     TENANT_ID = os.environ['SCHEMA_REGISTRY_AZURE_TENANT_ID']
     CLIENT_ID = os.environ['SCHEMA_REGISTRY_AZURE_CLIENT_ID']
     CLIENT_SECRET = os.environ['SCHEMA_REGISTRY_AZURE_CLIENT_SECRET']
     token_credential = ClientSecretCredential(TENANT_ID, CLIENT_ID, CLIENT_SECRET)
-    schema_registry_client = SchemaRegistryClient(endpoint=SCHEMA_REGISTRY_ENDPOINT, credential=token_credential)
+    schema_registry_client = SchemaRegistryClient(fully_qualified_namespace=SCHEMA_REGISTRY_FQN, credential=token_credential)
     return schema_registry_client, token_credential
 
 
 async def register_schema(schema_registry_client):
     # [START register_schema_async]
-    SCHEMA_GROUP = os.environ['SCHEMA_REGISTRY_GROUP']
-    SCHEMA_NAME = 'your-schema-name'
-    SERIALIZATION_TYPE = SerializationType.AVRO
-    SCHEMA_CONTENT = """{"namespace":"example.avro","type":"record","name":"User","fields":[{"name":"name","type":"string"},{"name":"favorite_number","type":["int","null"]},{"name":"favorite_color","type":["string","null"]}]}"""
-    schema_properties = await schema_registry_client.register_schema(SCHEMA_GROUP, SCHEMA_NAME, SCHEMA_CONTENT, SERIALIZATION_TYPE)
+    GROUP_NAME = os.environ['SCHEMA_REGISTRY_GROUP']
+    NAME = 'your-schema-name'
+    FORMAT = SchemaFormat.AVRO
+    SCHEMA_DEFINITION = """{"namespace":"example.avro","type":"record","name":"User","fields":[{"name":"name","type":"string"},{"name":"favorite_number","type":["int","null"]},{"name":"favorite_color","type":["string","null"]}]}"""
+    schema_properties = await schema_registry_client.register_schema(GROUP_NAME, NAME, SCHEMA_DEFINITION, FORMAT)
     schema_id = schema_properties.id
     # [END register_schema_async]
     return schema_id
 
 
-async def get_schema(schema_registry_client, schema_id):
+async def get_schema(schema_registry_client, id):
     # [START get_schema_async]
-    schema = await schema_registry_client.get_schema(schema_id)
-    schema_content = schema.content
+    schema = await schema_registry_client.get_schema(id)
+    schema_definition = schema.schema_definition
     # [END get_schema_async]
-    return schema_content
+    return schema_definition
 
 
 async def get_schema_id(schema_registry_client):
-    schema_group = os.environ['SCHEMA_REGISTRY_GROUP']
-    schema_name = 'your-schema-name'
-    serialization_type = SerializationType.AVRO
-    schema_content = """{"namespace":"example.avro","type":"record","name":"User","fields":[{"name":"name","type":"string"},{"name":"favorite_number","type":["int","null"]},{"name":"favorite_color","type":["string","null"]}]}"""
+    group_name = os.environ['SCHEMA_REGISTRY_GROUP']
+    name = 'your-schema-name'
+    format = SchemaFormat.AVRO
+    schema_definition = """{"namespace":"example.avro","type":"record","name":"User","fields":[{"name":"name","type":"string"},{"name":"favorite_number","type":["int","null"]},{"name":"favorite_color","type":["string","null"]}]}"""
 
     # [START get_schema_id_async]
-    schema_properties = await schema_registry_client.get_schema_properties(schema_group, schema_name, schema_content, serialization_type)
+    schema_properties = await schema_registry_client.get_schema_properties(group_name, name, schema_definition, format)
     schema_id = schema_properties.id
     # [END get_schema_id_async]
     return schema_id