+++ sdk/search/azure-search-documents/azure/search/documents/indexes/_search_indexer_client.py
@@ -15,7 +15,11 @@ from ._utils import (
     get_access_conditions,
     normalize_endpoint,
 )
-from .models import SearchIndexerDataSourceConnection
+from .models import (
+    EntityRecognitionSkillVersion,
+    SearchIndexerDataSourceConnection,
+    SentimentSkillVersion
+)
 from .._api_versions import DEFAULT_VERSION
 from .._headers_mixin import HeadersMixin
 from .._utils import get_authentication_policy
@@ -564,7 +568,9 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
 
         """
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
+        _validate_skillset(skillset)
         skillset = skillset._to_generated() if hasattr(skillset, '_to_generated') else skillset # pylint:disable=protected-access
+
         result = self._client.skillsets.create(skillset, **kwargs)
         return SearchIndexerSkillset._from_generated(result) # pylint:disable=protected-access
 
@@ -587,6 +593,7 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
             skillset, kwargs.pop("match_condition", MatchConditions.Unconditionally)
         )
         kwargs.update(access_condition)
+        _validate_skillset(skillset)
         skillset = skillset._to_generated() if hasattr(skillset, '_to_generated') else skillset # pylint:disable=protected-access
 
         result = self._client.skillsets.create_or_update(
@@ -596,3 +603,44 @@ class SearchIndexerClient(HeadersMixin):  # pylint: disable=R0904
             **kwargs
         )
         return SearchIndexerSkillset._from_generated(result) # pylint:disable=protected-access
+
+def _validate_skillset(skillset):
+    """Validates any multi-version skills in the skillset to verify that unsupported
+    parameters are not supplied by the user.
+    """
+    skills = getattr(skillset, 'skills', None)
+    if not skills:
+        return
+
+    error_strings = []
+    for skill in skills:
+        try:
+            skill_version = skill.get('skill_version')
+        except AttributeError:
+            skill_version = getattr(skill, 'skill_version', None)
+        if not skill_version:
+            continue
+
+        if skill_version == SentimentSkillVersion.V1:
+            unsupported = ['model_version', 'include_opinion_mining']
+        elif skill_version == SentimentSkillVersion.V3:
+            unsupported = []
+        elif skill_version == EntityRecognitionSkillVersion.V1:
+            unsupported = ['model_version']
+        elif skill_version == EntityRecognitionSkillVersion.V3:
+            unsupported = ['include_typeless_entities']
+
+        errors = []
+        for item in unsupported:
+            try:
+                if skill.get(item, None):
+                    errors.append(item)
+            except AttributeError:
+                if skill.__dict__.get(item, None):
+                    errors.append(item)
+        if errors:
+            error_strings.append("Unsupported parameters for skill version {}: {}".format(
+                skill_version, ", ".join(errors))
+            )
+    if error_strings:
+        raise ValueError("\n".join(error_strings))